🐛 fix: 기존 sunghwki가 사용하던 것을 옮긺, 또한 bigint가 적용되지 않는 문제를 해결함
