+++ packages/backend/src/stock/domain/stock.entity.ts
@@ -1,7 +1,13 @@
 import { Column, Entity, OneToMany, PrimaryColumn } from 'typeorm';
 import { DateEmbedded } from '@/common/dateEmbedded.entity';
 import { UserStock } from '@/stock/domain/userStock.entity';
-import { StockDaily } from './stockDaily.entity';
+import {
+  StockDaily,
+  StockMinutely,
+  StockMonthly,
+  StockWeekly,
+  StockYearly,
+} from './stockData.entity';
 
 @Entity()
 export class Stock {
@@ -26,6 +32,18 @@ export class Stock {
   @OneToMany(() => UserStock, (userStock) => userStock.stock)
   userStocks?: UserStock[];
 
+  @OneToMany(() => StockMinutely, (stockMinutely) => stockMinutely.stock)
+  stockMinutely?: StockMinutely[];
+
   @OneToMany(() => StockDaily, (stockDaily) => stockDaily.stock)
   stockDaily?: StockDaily[];
+
+  @OneToMany(() => StockWeekly, (stockWeekly) => stockWeekly.stock)
+  stockWeekly?: StockWeekly[];
+
+  @OneToMany(() => StockMonthly, (stockMonthly) => stockMonthly.stock)
+  stockMonthly?: StockMonthly[];
+
+  @OneToMany(() => StockYearly, (stockYearly) => stockYearly.stock)
+  stockYearly?: StockYearly[];
 }
