--- packages/backend/src/stock/domain/stockDaily.entity.ts
@@ -1,6 +1,0 @@
-import { Entity } from 'typeorm';
-import { StockPeriod } from './stockPeriod';
-
-@Entity('stock_daily')
-export class StockDaily extends StockPeriod {
-}
