--- packages/backend/src/stock/domain/stockMinutely.entity.ts
@@ -1,6 +1,0 @@
-import { Entity } from 'typeorm';
-import { StockPeriod } from './stockPeriod';
-
-@Entity('stock_minutely')
-export class StockMinutely extends StockPeriod {
-}
