--- packages/backend/src/stock/domain/stockMonthly.entity.ts
@@ -1,6 +1,0 @@
-import { Entity } from 'typeorm';
-import { StockPeriod } from './stockPeriod';
-
-@Entity('stock_monthly')
-export class StockMonthly extends StockPeriod {
-}
