--- packages/backend/src/stock/domain/stockWeekly.entity.ts
@@ -1,6 +1,0 @@
-import { Entity } from 'typeorm';
-import { StockPeriod } from './stockPeriod';
-
-@Entity('stock_weekly')
-export class StockWeekly extends StockPeriod {
-}
