--- packages/backend/src/stock/domain/stockYearly.entity.ts
@@ -1,6 +1,0 @@
-import { Entity } from 'typeorm';
-import { StockPeriod } from './stockPeriod';
-
-@Entity('stock_yearly')
-export class StockYearly extends StockPeriod {
-}
